Move PartialResults to separate file
